requirements.txt
requests
pydub
pyyaml